move validateRanges out of utils
